requirements.txt
Flask[async]==3.1.1
flask-cors==6.0.1
requests==2.32.4
googletrans==4.0.2
gunicorn==21.2.0
python-dotenv==1.1.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
requests==2.32.4
flask-bcrypt
flask-jwt-extended
Flask-Mail==0.10.0
Flask-Limiter==3.12
PyPDF2==3.0.1
python-docx==1.2.0

